feat: add useClickOutside
