+++ src/useRootClose.ts
@@ -1,36 +1,20 @@
-import contains from 'dom-helpers/contains';
 import listen from 'dom-helpers/listen';
 import ownerDocument from 'dom-helpers/ownerDocument';
-import { useCallback, useEffect, useRef } from 'react';
+import { useEffect } from 'react';
 
 import useEventCallback from '@restart/hooks/useEventCallback';
-import warning from 'warning';
+import useClickOutside, {
+  ClickOutsideOptions,
+  getRefTarget,
+} from './useClickOutside';
 
 const escapeKeyCode = 27;
 const noop = () => {};
 
-export type MouseEvents = {
-  [K in keyof GlobalEventHandlersEventMap]: GlobalEventHandlersEventMap[K] extends MouseEvent
-    ? K
-    : never;
-}[keyof GlobalEventHandlersEventMap];
-
-function isLeftClickEvent(event: MouseEvent) {
-  return event.button === 0;
-}
-
-function isModifiedEvent(event: MouseEvent) {
-  return !!(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey);
-}
-
-const getRefTarget = (
-  ref: React.RefObject<Element> | Element | null | undefined,
-) => ref && ('current' in ref ? ref.current : ref);
-
-export interface RootCloseOptions {
+export interface RootCloseOptions extends ClickOutsideOptions {
   disabled?: boolean;
-  clickTrigger?: MouseEvents;
 }
+
 /**
  * The `useRootClose` hook registers your callback on the document
  * when rendered. Powers the `<Overlay/>` component. This is used achieve modal
@@ -46,35 +30,11 @@ export interface RootCloseOptions {
 function useRootClose(
   ref: React.RefObject<Element> | Element | null | undefined,
   onRootClose: (e: Event) => void,
-  { disabled, clickTrigger = 'click' }: RootCloseOptions = {},
+  { disabled, clickTrigger }: RootCloseOptions = {},
 ) {
-  const preventMouseRootCloseRef = useRef(false);
   const onClose = onRootClose || noop;
 
-  const handleMouseCapture = useCallback(
-    (e) => {
-      const currentTarget = getRefTarget(ref);
-
-      warning(
-        !!currentTarget,
-        'RootClose captured a close event but does not have a ref to compare it to. ' +
-          'useRootClose(), should be passed a ref that resolves to a DOM node',
-      );
-
-      preventMouseRootCloseRef.current =
-        !currentTarget ||
-        isModifiedEvent(e) ||
-        !isLeftClickEvent(e) ||
-        !!contains(currentTarget, e.target);
-    },
-    [ref],
-  );
-
-  const handleMouse = useEventCallback((e: MouseEvent) => {
-    if (!preventMouseRootCloseRef.current) {
-      onClose(e);
-    }
-  });
+  useClickOutside(ref, onClose, { disabled, clickTrigger });
 
   const handleKeyUp = useEventCallback((e: KeyboardEvent) => {
     if (e.keyCode === escapeKeyCode) {
@@ -91,25 +51,6 @@ function useRootClose(
     // https://github.com/facebook/react/issues/20074
     let currentEvent = (doc.defaultView || window).event;
 
-    // Use capture for this listener so it fires before React's listener, to
-    // avoid false positives in the contains() check below if the target DOM
-    // element is removed in the React mouse callback.
-    const removeMouseCaptureListener = listen(
-      doc as any,
-      clickTrigger,
-      handleMouseCapture,
-      true,
-    );
-
-    const removeMouseListener = listen(doc as any, clickTrigger, (e) => {
-      // skip if this event is the same as the one running when we added the handlers
-      if (e === currentEvent) {
-        currentEvent = undefined;
-        return;
-      }
-      handleMouse(e);
-    });
-
     const removeKeyupListener = listen(doc as any, 'keyup', (e) => {
       // skip if this event is the same as the one running when we added the handlers
       if (e === currentEvent) {
@@ -119,27 +60,10 @@ function useRootClose(
       handleKeyUp(e);
     });
 
-    let mobileSafariHackListeners = [] as Array<() => void>;
-    if ('ontouchstart' in doc.documentElement) {
-      mobileSafariHackListeners = [].slice
-        .call(doc.body.children)
-        .map((el) => listen(el, 'mousemove', noop));
-    }
-
     return () => {
-      removeMouseCaptureListener();
-      removeMouseListener();
       removeKeyupListener();
-      mobileSafariHackListeners.forEach((remove) => remove());
     };
-  }, [
-    ref,
-    disabled,
-    clickTrigger,
-    handleMouseCapture,
-    handleMouse,
-    handleKeyUp,
-  ]);
+  }, [ref, disabled, handleKeyUp]);
 }
 
 export default useRootClose;
